requirements.txt
# OriginX - 图像/视频质量诊断系统
# 核心依赖

# 图像处理
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
scipy>=1.7.0

# API框架
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# CLI
click>=8.0.0
rich>=13.0.0

# 配置
pyyaml>=6.0
python-dotenv>=1.0.0

# 工具
requests>=2.28.0
aiofiles>=23.0.0
httpx>=0.24.0

# 定时任务 (V1.5)
apscheduler>=3.10.0

# 报告生成 (V1.5)
openpyxl>=3.1.0
reportlab>=4.0.0
jinja2>=3.1.0

